Moved raw signing to hazmat module
