+++ src/signing.rs
@@ -810,7 +810,7 @@ impl ExpandedSecretKey {
 }
 
 //
-// Signing funcitons. These are private to this module unless explicitly exported.
+// Signing functions. These are pub(crate) so that the hazmat can use them
 //
 
 /// The plain, non-prehashed, signing function for Ed25519. `scalar` is the secret scalar of the
@@ -820,7 +820,7 @@ impl ExpandedSecretKey {
 /// See docs on `raw_sign()` for more detail.
 #[allow(non_snake_case)]
 #[inline(always)]
-fn private_raw_sign(
+pub(crate) fn private_raw_sign(
     scalar: Scalar,
     hash_prefix: [u8; 32],
     message: &[u8],
@@ -852,7 +852,7 @@ fn private_raw_sign(
 #[cfg(feature = "digest")]
 #[allow(non_snake_case)]
 #[inline(always)]
-fn private_raw_sign_prehashed<'a, D>(
+pub(crate) fn private_raw_sign_prehashed<'a, D>(
     scalar: Scalar,
     hash_prefix: [u8; 32],
     prehashed_message: D,
@@ -912,83 +912,4 @@ where
     let s: Scalar = (k * scalar) + r;
 
     Ok(InternalSignature { R, s }.into())
-}
-
-//
-// Exports
-//
-
-/// The plain, non-prehashed, signing function for Ed25519.
-///
-/// **Unsafe:** Do NOT use this function unless you absolutely must. Misuse of this function can
-/// expose your private key. See [here](https://github.com/MystenLabs/ed25519-unsafe-libs) for more
-/// details on this attack.
-///
-/// # Inputs
-///
-/// * `scalar` is the secret scalar of the signing key
-/// * `hash_prefix` is the domain separator that, along with the message itself, is used to
-///   deterministically generate the `R` part of the signature.
-///
-/// `scalar` and `hash_prefix` are usually selected such that `scalar || hash_prefix = H(sk)` where
-/// `sk` is the signing key
-#[cfg(feature = "raw-sign")]
-fn raw_sign(
-    scalar: Scalar,
-    hash_prefix: [u8; 32],
-    message: &[u8],
-    verifying_key: &VerifyingKey,
-) -> Signature {
-    private_raw_sign(scalar, hash_prefix, message, verifying_key)
-}
-
-/// Sign a `prehashed_message` with this `ExpandedSecretKey` using the Ed25519ph algorithm defined
-/// in [RFC8032 §5.1][rfc8032].
-///
-/// **Unsafe:** Do NOT use this function unless you absolutely must. Misuse of this function can
-/// expose your private key. See [here](https://github.com/MystenLabs/ed25519-unsafe-libs) for more
-/// details on this attack.
-///
-/// # Inputs
-///
-/// * `scalar` is the secret scalar of the signing key
-/// * `hash_prefix` is the domain separator that, along with the prehashed message, is used to
-///   deterministically generate the `R` part of the signature.
-/// * `prehashed_message` is an instantiated hash digest with 512-bits of
-///   output which has had the message to be signed previously fed into its
-///   state.
-/// * `verifying_key` is a [`VerifyingKey`] which corresponds to this secret key.
-/// * `context` is an optional context string, up to 255 bytes inclusive,
-///   which may be used to provide additional domain separation.  If not
-///   set, this will default to an empty string.
-///
-/// `scalar` and `hash_prefix` are usually selected such that `scalar || hash_prefix = H(sk)` where
-/// `sk` is the signing key
-///
-/// # Returns
-///
-/// A `Result` whose `Ok` value is an Ed25519ph [`Signature`] on the
-/// `prehashed_message` if the context was 255 bytes or less, otherwise
-/// a `SignatureError`.
-///
-/// [rfc8032]: https://tools.ietf.org/html/rfc8032#section-5.1
-#[cfg(all(feature = "digest", feature = "raw-sign"))]
-#[allow(non_snake_case)]
-pub fn raw_sign_prehashed<'a, D>(
-    scalar: Scalar,
-    hash_prefix: [u8; 32],
-    prehashed_message: D,
-    verifying_key: &VerifyingKey,
-    context: Option<&'a [u8]>,
-) -> Result<Signature, SignatureError>
-where
-    D: Digest<OutputSize = U64>,
-{
-    private_raw_sign_prehashed(
-        scalar,
-        hash_prefix,
-        prehashed_message,
-        verifying_key,
-        context,
-    )
 }
